split status_monitor.py
